--- client/src/App/App.js
@@ -1,14 +1,0 @@
-import Main from '../components/Main/Main';
-import NavBar from '../components/NavBar/NavBar';
-import './App.css';
-
-function App() {
-	return (
-		<div className="App">
-			<NavBar />
-			<Main />
-		</div>
-	);
-}
-
-export default App;
