+++ client/src/components/About.js
@@ -1,0 +1,9 @@
+function About() {
+  return (
+    <div>
+      <h1>About Page</h1>
+    </div>
+  );
+}
+
+export default About;
