+++ client/src/components/App.js
@@ -1,0 +1,14 @@
+import Main from './Main';
+import NavBar from './NavBar';
+import '../stylesheets/App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <NavBar />
+      <Main />
+    </div>
+  );
+}
+
+export default App;
