--- client/src/components/NavBar/NavBar.scss
@@ -1,6 +1,0 @@
-ul{
-
-  padding: 0px;
-  list-style-type: none;
-}
-
