+++ client/src/index.js
@@ -2,9 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
-
-
-import App from './App/App';
+import App from './components/App';
 
 ReactDOM.render(
   <React.StrictMode>
